requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
streamlit
torch
torchvision
numpy
pillow
opencv-python
pandas
matplotlib